refactor(xray-lite): move SubsegmentSession to session
Moves `SubsegmentSession` to a new submodule `session`.
